assignmenttexter-refactor: AssignmentTexter renames based on previous moves
