+++ src/components/AssignmentTexter/ContactController.jsx
@@ -1,11 +1,11 @@
 import PropTypes from "prop-types";
 import React from "react";
 import IconButton from "material-ui/IconButton/IconButton";
-import LoadingIndicator from "./LoadingIndicator";
+import LoadingIndicator from "../LoadingIndicator";
 import { StyleSheet, css } from "aphrodite";
 import { withRouter } from "react-router";
 import Check from "material-ui/svg-icons/action/check-circle";
-import Empty from "../components/Empty";
+import Empty from "../Empty";
 import RaisedButton from "material-ui/RaisedButton";
 
 const styles = StyleSheet.create({
@@ -23,7 +23,7 @@ const styles = StyleSheet.create({
   }
 });
 
-export class AssignmentTexter extends React.Component {
+export class ContactController extends React.Component {
   constructor(props) {
     super(props);
 
@@ -445,7 +445,7 @@ export class AssignmentTexter extends React.Component {
   }
 }
 
-AssignmentTexter.propTypes = {
+ContactController.propTypes = {
   assignment: PropTypes.object, // current assignment
   contacts: PropTypes.array, // contacts for current assignment
   allContactsCount: PropTypes.number,
@@ -459,4 +459,4 @@ AssignmentTexter.propTypes = {
   messageStatusFilter: PropTypes.string
 };
 
-export default withRouter(AssignmentTexter);
+export default withRouter(ContactController);
