+++ src/components/AssignmentTexter/OldToolbar.jsx
@@ -1,9 +1,9 @@
 import PropTypes from "prop-types";
 import React from "react";
 import { Toolbar, ToolbarGroup, ToolbarTitle } from "material-ui/Toolbar";
-import { getDisplayPhoneNumber } from "../lib/phone-format";
-import { getLocalTime, getContactTimezone } from "../lib/timezones";
-import { getProcessEnvDstReferenceTimezone } from "../lib/tz-helpers";
+import { getDisplayPhoneNumber } from "../../lib/phone-format";
+import { getLocalTime, getContactTimezone } from "../../lib/timezones";
+import { getProcessEnvDstReferenceTimezone } from "../../lib/tz-helpers";
 import { grey100 } from "material-ui/styles/colors";
 
 const inlineStyles = {
