+++ src/components/AssignmentTexter/Toolbar.jsx
@@ -2,9 +2,9 @@ import PropTypes from "prop-types";
 import React from "react";
 import { StyleSheet, css } from "aphrodite";
 import { Toolbar, ToolbarGroup, ToolbarTitle } from "material-ui/Toolbar";
-import { getDisplayPhoneNumber } from "../lib/phone-format";
-import { getLocalTime, getContactTimezone } from "../lib/timezones";
-import { getProcessEnvDstReferenceTimezone } from "../lib/tz-helpers";
+import { getDisplayPhoneNumber } from "../../lib/phone-format";
+import { getLocalTime, getContactTimezone } from "../../lib/timezones";
+import { getProcessEnvDstReferenceTimezone } from "../../lib/tz-helpers";
 import ActionFace from "material-ui/svg-icons/action/face";
 import IconButton from "material-ui/IconButton/IconButton";
 import { grey100 } from "material-ui/styles/colors";
